API 70% done

+++ stock-market/Model/BaseStock.cs
@@ -6,7 +6,7 @@ namespace stock_market.Model
         public int id { get; set; }
         public string ticker { get; set; }
         public string name { get; set; }
-        public int current_price { get; set; }
+        public double current_price { get; set; }
     }
 }
