+++ stock-market/Model/User.cs
@@ -12,9 +12,9 @@ namespace stock_market.Model
         public int id { get; set; }
         public string first_name { get; set; }
         public string last_name { get; set; }
-        public int curr_balance { get; set; }
-        public int curr_balance_liquid { get; set; }
-        public int curr_balance_stock { get; set; }
+        public double curr_balance { get; set; }
+        public double curr_balance_liquid { get; set; }
+        public double curr_balance_stock { get; set; }
 
         virtual public List<Transaction> transactions { get; set; }
     }
